+++ BTrees/BTreeModuleTemplate.c
@@ -53,8 +53,8 @@
  */
 #define MODULE_NAME "BTrees." MOD_NAME_PREFIX "BTree."
 
-static PyObject *sort_str, *reverse_str, *__setstate___str,
-    *_bucket_type_str;
+static PyObject *sort_str, *reverse_str, *__setstate___str;
+static PyObject *_bucket_type_str, *max_internal_size_str, *max_leaf_size_str;
 static PyObject *ConflictError = NULL;
 
 static void PyVar_Assign(PyObject **v, PyObject *e) { Py_XDECREF(*v); *v=e;}
@@ -63,8 +63,6 @@ static void PyVar_Assign(PyObject **v, PyObject *e) { Py_XDECREF(*v); *v=e;}
 #define OBJECT(O) ((PyObject*)(O))
 
 #define MIN_BUCKET_ALLOC 16
-#define MAX_BTREE_SIZE(B) DEFAULT_MAX_BTREE_SIZE
-#define MAX_BUCKET_SIZE(B) DEFAULT_MAX_BUCKET_SIZE
 
 #define SameType_Check(O1, O2) (Py_TYPE((O1))==Py_TYPE((O2)))
 
@@ -223,6 +221,8 @@ typedef struct BTree_s {
    * data[len].key is positive infinity.
    */
   BTreeItem *data;
+  long max_internal_size;
+  long max_leaf_size;
 } BTree;
 
 static PyTypeObject BTreeType;
@@ -363,21 +363,11 @@ PreviousBucket(Bucket **current, Bucket *first)
 	PER_USE_OR_RETURN(first, -1);
         first = first->next;
 
+    	((trailing)->state==cPersistent_STICKY_STATE
+     	&&
+     	((trailing)->state=cPersistent_UPTODATE_STATE));
 
-
-
-
-
-
-    ((trailing)->state==cPersistent_STICKY_STATE
-     &&
-     ((trailing)->state=cPersistent_UPTODATE_STATE));
-
-    PER_ACCESSED(trailing);
-
-
-
-
+    	PER_ACCESSED(trailing);
 
 	if (first == *current) {
 	    *current = trailing;
@@ -540,6 +530,13 @@ module_init(void)
         return NULL;
     _bucket_type_str = INTERN("_bucket_type");
     if (!_bucket_type_str)
+        return NULL;
+
+    max_internal_size_str = INTERN("max_internal_size");
+    if (! max_internal_size_str)
+        return NULL;
+    max_leaf_size_str = INTERN("max_leaf_size");
+    if (! max_leaf_size_str)
         return NULL;
 
     /* Grab the ConflictError class */