+++ BTrees/IFBTree.py
@@ -14,7 +14,7 @@
 
 __all__ = ('Bucket', 'Set', 'BTree', 'TreeSet',
            'IFBucket', 'IFSet', 'IFBTree', 'IFTreeSet',
-           'union', 'intersection', 'difference',  
+           'union', 'intersection', 'difference',
            'weightedUnion', 'weightedIntersection', 'multiunion',
           )
 
@@ -38,13 +38,13 @@ from ._base import to_float as _to_value
 from ._base import union as _union
 from ._base import weightedIntersection as _weightedIntersection
 from ._base import weightedUnion as _weightedUnion
+from ._base import _fix_pickle
 
 _BUCKET_SIZE = 120
 _TREE_SIZE = 500
 using64bits = False
 
 class IFBucketPy(Bucket):
-    MAX_SIZE = _BUCKET_SIZE
     _to_key = _to_key
     _to_value = _to_value
     MERGE = MERGE
@@ -53,7 +53,6 @@ class IFBucketPy(Bucket):
 
 
 class IFSetPy(Set):
-    MAX_SIZE = _BUCKET_SIZE
     _to_key = _to_key
     MERGE = MERGE
     MERGE_WEIGHT = MERGE_WEIGHT_numeric
@@ -61,7 +60,8 @@ class IFSetPy(Set):
 
 
 class IFBTreePy(BTree):
-    MAX_SIZE = _TREE_SIZE
+    max_leaf_size = _BUCKET_SIZE
+    max_internal_size = _TREE_SIZE
     _to_key = _to_key
     _to_value = _to_value
     MERGE = MERGE
@@ -70,7 +70,8 @@ class IFBTreePy(BTree):
 
 
 class IFTreeSetPy(TreeSet):
-    MAX_SIZE = _TREE_SIZE
+    max_leaf_size = _BUCKET_SIZE
+    max_internal_size = _TREE_SIZE
     _to_key = _to_key
     MERGE = MERGE
     MERGE_WEIGHT = MERGE_WEIGHT_numeric
@@ -133,5 +134,7 @@ Bucket = IFBucket
 Set = IFSet
 BTree = IFBTree
 TreeSet = IFTreeSet
+
+_fix_pickle(globals(), __name__)
 
 moduleProvides(IIntegerFloatBTreeModule)