+++ BTrees/LLBTree.py
@@ -14,7 +14,7 @@
 
 __all__ = ('Bucket', 'Set', 'BTree', 'TreeSet',
            'LLBucket', 'LLSet', 'LLBTree', 'LLTreeSet',
-           'union', 'intersection', 'difference',  
+           'union', 'intersection', 'difference',
            'weightedUnion', 'weightedIntersection', 'multiunion',
           )
 
@@ -38,6 +38,7 @@ from ._base import to_long as _to_value
 from ._base import union as _union
 from ._base import weightedIntersection as _weightedIntersection
 from ._base import weightedUnion as _weightedUnion
+from ._base import _fix_pickle
 
 _BUCKET_SIZE = 120
 _TREE_SIZE = 500
@@ -45,7 +46,6 @@ using64bits = True
 
 
 class LLBucketPy(Bucket):
-    MAX_SIZE = _BUCKET_SIZE
     _to_key = _to_key
     _to_value = _to_value
     MERGE = MERGE
@@ -54,7 +54,6 @@ class LLBucketPy(Bucket):
 
 
 class LLSetPy(Set):
-    MAX_SIZE = _BUCKET_SIZE
     _to_key = _to_key
     MERGE = MERGE
     MERGE_WEIGHT = MERGE_WEIGHT_numeric
@@ -62,7 +61,8 @@ class LLSetPy(Set):
 
 
 class LLBTreePy(BTree):
-    MAX_SIZE = _TREE_SIZE
+    max_leaf_size = _BUCKET_SIZE
+    max_internal_size = _TREE_SIZE
     _to_key = _to_key
     _to_value = _to_value
     MERGE = MERGE
@@ -71,7 +71,8 @@ class LLBTreePy(BTree):
 
 
 class LLTreeSetPy(TreeSet):
-    MAX_SIZE = _TREE_SIZE
+    max_leaf_size = _BUCKET_SIZE
+    max_internal_size = _TREE_SIZE
     _to_key = _to_key
     MERGE = MERGE
     MERGE_WEIGHT = MERGE_WEIGHT_numeric
@@ -134,5 +135,7 @@ Bucket = LLBucket
 Set = LLSet
 BTree = LLBTree
 TreeSet = LLTreeSet
+
+_fix_pickle(globals(), __name__)
 
 moduleProvides(IIntegerIntegerBTreeModule)