+++ BTrees/OIBTree.py
@@ -14,7 +14,7 @@
 
 __all__ = ('Bucket', 'Set', 'BTree', 'TreeSet',
            'OIBucket', 'OISet', 'OIBTree', 'OITreeSet',
-           'union', 'intersection', 'difference',  
+           'union', 'intersection', 'difference',
            'weightedUnion', 'weightedIntersection',
           )
 
@@ -37,13 +37,13 @@ from ._base import to_int as _to_value
 from ._base import union as _union
 from ._base import weightedIntersection as _weightedIntersection
 from ._base import weightedUnion as _weightedUnion
+from ._base import _fix_pickle
 
 _BUCKET_SIZE = 60
 _TREE_SIZE = 250
 using64bits = True
 
 class OIBucketPy(Bucket):
-    MAX_SIZE = _BUCKET_SIZE
     _to_key = _to_key
     _to_value = _to_value
     MERGE = MERGE
@@ -52,7 +52,6 @@ class OIBucketPy(Bucket):
 
 
 class OISetPy(Set):
-    MAX_SIZE = _BUCKET_SIZE
     _to_key = _to_key
     MERGE = MERGE
     MERGE_WEIGHT = MERGE_WEIGHT_numeric
@@ -60,7 +59,8 @@ class OISetPy(Set):
 
 
 class OIBTreePy(BTree):
-    MAX_SIZE = _TREE_SIZE
+    max_leaf_size = _BUCKET_SIZE
+    max_internal_size = _TREE_SIZE
     _to_key = _to_key
     _to_value = _to_value
     MERGE = MERGE
@@ -69,7 +69,8 @@ class OIBTreePy(BTree):
 
 
 class OITreeSetPy(TreeSet):
-    MAX_SIZE = _TREE_SIZE
+    max_leaf_size = _BUCKET_SIZE
+    max_internal_size = _TREE_SIZE
     _to_key = _to_key
     MERGE = MERGE
     MERGE_WEIGHT = MERGE_WEIGHT_numeric
@@ -130,5 +131,7 @@ Bucket = OIBucket
 Set = OISet
 BTree = OIBTree
 TreeSet = OITreeSet
+
+_fix_pickle(globals(), __name__)
 
 moduleProvides(IObjectIntegerBTreeModule)