+++ BTrees/tests/test__base.py
@@ -30,6 +30,8 @@ class Test_Base(unittest.TestCase):
 
     def _makeOne(self, items=None):
         class _Test(self._getTargetClass()):
+            max_leaf_size = 10
+            max_internal_size = 15
             def clear(self):
                 self._data = {}
             def update(self, d):
@@ -595,6 +597,19 @@ class BucketTests(unittest.TestCase):
         self.assertTrue(bucket._next is new_b)
         self.assertTrue(new_b._next is next_b)
 
+    def test__split_filled_explicit_index(self):
+        bucket = self._makeOne()
+        next_b = bucket._next = self._makeOne()
+        for i, c in enumerate('abcdef'):
+            bucket[c] = i
+        new_b = bucket._split(2)
+        self.assertEqual(list(bucket._keys), ['a', 'b'])
+        self.assertEqual(list(bucket._values), [0, 1])
+        self.assertEqual(list(new_b._keys), ['c', 'd', 'e', 'f'])
+        self.assertEqual(list(new_b._values), [2, 3, 4, 5])
+        self.assertTrue(bucket._next is new_b)
+        self.assertTrue(new_b._next is next_b)
+
     def test_keys_empty_no_args(self):
         bucket = self._makeOne()
         self.assertEqual(bucket.keys(), [])
@@ -762,6 +777,17 @@ class BucketTests(unittest.TestCase):
         self.assertEqual(bucket.items(), EXPECTED)
         self.assertTrue(bucket._next is next_b)
 
+    def test__p_resolveConflict_x_on_com_next_old_new_None(self):
+        from ..Interfaces import BTreesConflictError
+        bucket = self._makeOne()
+        N_NEW = object()
+        s_old = None
+        s_com = ((), N_NEW)
+        s_new = None
+        e = self.assertRaises(BTreesConflictError,
+                              bucket._p_resolveConflict, s_old, s_com, s_new)
+        self.assertEqual(e.reason, 0)
+
     def test__p_resolveConflict_x_on_com_next(self):
         from ..Interfaces import BTreesConflictError
         bucket = self._makeOne()
@@ -769,6 +795,17 @@ class BucketTests(unittest.TestCase):
         s_old = ((), None)
         s_com = ((), N_NEW)
         s_new = ((), None)
+        e = self.assertRaises(BTreesConflictError,
+                              bucket._p_resolveConflict, s_old, s_com, s_new)
+        self.assertEqual(e.reason, 0)
+
+    def test__p_resolveConflict_x_on_new_next_old_com_None(self):
+        from ..Interfaces import BTreesConflictError
+        bucket = self._makeOne()
+        N_NEW = object()
+        s_old = None
+        s_com = None
+        s_new = ((), N_NEW)
         e = self.assertRaises(BTreesConflictError,
                               bucket._p_resolveConflict, s_old, s_com, s_new)
         self.assertEqual(e.reason, 0)
@@ -1149,6 +1186,28 @@ class SetTests(unittest.TestCase):
         self.assertTrue(_set._next is new_b)
         self.assertTrue(new_b._next is next_b)
 
+    def test__split_filled_explicit_index(self):
+        _set = self._makeOne()
+        next_b = _set._next = self._makeOne()
+        for c in 'abcdef':
+            _set.add(c)
+        new_b = _set._split(2)
+        self.assertEqual(list(_set._keys), ['a', 'b'])
+        self.assertEqual(list(new_b._keys), ['c', 'd', 'e', 'f'])
+        self.assertTrue(_set._next is new_b)
+        self.assertTrue(new_b._next is next_b)
+
+    def test__p_resolveConflict_x_on_com_next_old_new_None(self):
+        from ..Interfaces import BTreesConflictError
+        _set = self._makeOne()
+        N_NEW = object()
+        s_old = None
+        s_com = ((), N_NEW)
+        s_new = None
+        e = self.assertRaises(BTreesConflictError,
+                              _set._p_resolveConflict, s_old, s_com, s_new)
+        self.assertEqual(e.reason, 0)
+
     def test__p_resolveConflict_x_on_com_next(self):
         from ..Interfaces import BTreesConflictError
         _set = self._makeOne()
@@ -1156,6 +1215,17 @@ class SetTests(unittest.TestCase):
         s_old = ((), None)
         s_com = ((), N_NEW)
         s_new = ((), None)
+        e = self.assertRaises(BTreesConflictError,
+                              _set._p_resolveConflict, s_old, s_com, s_new)
+        self.assertEqual(e.reason, 0)
+
+    def test__p_resolveConflict_x_on_new_next_old_com_None(self):
+        from ..Interfaces import BTreesConflictError
+        _set = self._makeOne()
+        N_NEW = object()
+        s_old = None
+        s_com = None
+        s_new = ((), N_NEW)
         e = self.assertRaises(BTreesConflictError,
                               _set._p_resolveConflict, s_old, s_com, s_new)
         self.assertEqual(e.reason, 0)
@@ -1396,13 +1466,13 @@ class Test_Tree(unittest.TestCase):
     def _makeOne(self, items=None):
         from .._base import Bucket
         class _Bucket(Bucket):
-            MAX_SIZE = 10
             def _to_key(self, k):
                 return k
         class _Test(self._getTargetClass()):
             _to_key = _to_value = lambda self, x: x
             _bucket_type = _Bucket
-            MAX_SIZE = 15
+            max_leaf_size = 10
+            max_internal_size = 15
         return _Test(items)
 
     def test_setdefault_miss(self):
@@ -1829,7 +1899,7 @@ class Test_Tree(unittest.TestCase):
             tree[key] = i
             KEYS.append(key)
         fb = tree._firstbucket
-        new_t = tree._split()
+        new_t = tree._split(tree.max_internal_size - 2)
         # Note that original tree still links to split buckets
         self.assertEqual(list(tree), KEYS)
         self.assertTrue(tree._firstbucket is fb)
@@ -2135,7 +2205,6 @@ class Test_TreeItems(unittest.TestCase):
     def _makeBucket(self, items=None):
         from .._base import Bucket
         class _Bucket(Bucket):
-            MAX_SIZE = 10
             def _to_key(self, k):
                 return k
         return _Bucket(items)
@@ -2214,13 +2283,13 @@ class TreeTests(unittest.TestCase):
     def _makeOne(self, items=None):
         from .._base import Bucket
         class _Bucket(Bucket):
-            MAX_SIZE = 10
             def _to_key(self, k):
                 return k
         class _Test(self._getTargetClass()):
             _to_key = _to_value = lambda self, x: x
             _bucket_type = _Bucket
-            MAX_SIZE = 15
+            max_leaf_size = 10
+            max_internal_size = 15
         return _Test(items)
 
     def test_get_empty_miss(self):
@@ -2359,13 +2428,13 @@ class TreeSetTests(unittest.TestCase):
     def _makeOne(self, items=None):
         from .._base import Bucket
         class _Bucket(Bucket):
-            MAX_SIZE = 10
             def _to_key(self, k):
                 return k
         class _Test(self._getTargetClass()):
             _to_key = _to_value = lambda self, x: x
             _bucket_type = _Bucket
-            MAX_SIZE = 15
+            max_leaf_size = 10
+            max_internal_size = 15
         return _Test(items)
 
     def test_add_new_key(self):